Readme.md
Image editing or processing from one image to fourth quadrant image which is the same size as the original image. By resizing and adding some decoration on the image
by using Opencv  and numpy in python. you can see the original image and the fourth quadrant resizing and decorated image.